refactor : pay 도메인 분리
