--- blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/entity/PayStatusEnum.java
@@ -1,5 +1,0 @@
-package com.sparta.blackwhitedeliverydriver.entity;
-
-public enum PayStatusEnum {
-    SUCCESS, REFUND;
-}
