+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/pay/dto/Amount.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.pay.dto;
 
 import lombok.Getter;
 import lombok.ToString;
