+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/pay/service/PayService.java
@@ -1,36 +1,35 @@
-package com.sparta.blackwhitedeliverydriver.service;
+package com.sparta.blackwhitedeliverydriver.pay.service;
 
-import com.sparta.blackwhitedeliverydriver.dto.PayApproveResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayCancelResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayGetDetailResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayGetResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayReadyResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayRefundRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayRefundResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.PayRequestDto;
-import com.sparta.blackwhitedeliverydriver.entity.Order;
-import com.sparta.blackwhitedeliverydriver.entity.OrderProduct;
-import com.sparta.blackwhitedeliverydriver.entity.OrderStatusEnum;
-import com.sparta.blackwhitedeliverydriver.entity.OrderTypeEnum;
-import com.sparta.blackwhitedeliverydriver.entity.Pay;
-import com.sparta.blackwhitedeliverydriver.entity.PayStatusEnum;
-import com.sparta.blackwhitedeliverydriver.entity.User;
-import com.sparta.blackwhitedeliverydriver.entity.UserRoleEnum;
-import com.sparta.blackwhitedeliverydriver.exception.ExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.OrderExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.PayExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.repository.OrderProductRepository;
-import com.sparta.blackwhitedeliverydriver.repository.OrderRepository;
-import com.sparta.blackwhitedeliverydriver.repository.PayRepository;
-import com.sparta.blackwhitedeliverydriver.repository.UserRepository;
-import com.sparta.blackwhitedeliverydriver.util.HttpUtil;
-import com.sparta.blackwhitedeliverydriver.util.PayUtil;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayApproveResponseDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayCancelResponseDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayGetDetailResponseDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayGetResponseDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayReadyResponseDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayRefundRequestDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayRefundResponseDto;
+import com.sparta.blackwhitedeliverydriver.pay.dto.PayRequestDto;
+import com.sparta.blackwhitedeliverydriver.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderProduct;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.order.entity.OrderTypeEnum;
+import com.sparta.blackwhitedeliverydriver.pay.entity.Pay;
+import com.sparta.blackwhitedeliverydriver.pay.entity.PayStatusEnum;
+import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.order.exception.OrderExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.pay.exception.PayExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.order.repository.OrderProductRepository;
+import com.sparta.blackwhitedeliverydriver.order.repository.OrderRepository;
+import com.sparta.blackwhitedeliverydriver.pay.repository.PayRepository;
+import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.pay.util.HttpUtil;
+import com.sparta.blackwhitedeliverydriver.pay.util.PayUtil;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
